# packages/shared-types/openapi.yaml
openapi: 3.0.3
info:
  title: TKAPrep API
  version: 0.1.0
  description: |
    REST API for the TKAPrep platform.

    All authenticated endpoints require `Authorization: Bearer <access_token>`.

    See `spec.md` in the repo root for the full product specification.
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      operationId: postAuthRegister
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags: [Auth]
      summary: Authenticate and receive tokens
      operationId: postAuthLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Rotate refresh token and get a new access token
      operationId: postAuthRefresh
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Tokens refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Invalid input
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  issue:
                    type: string
...
    RegisterRequest:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
          example: "Budi Santoso"
        email:
          type: string
          format: email
          example: "budi@example.com"
        password:
          type: string
          minLength: 8
          maxLength: 72
          example: "s3cur3P@ss"
        role:
          type: string
          enum: [student, contributor]
          example: student
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: "budi@example.com"
        password:
          type: string
          example: "s3cur3P@ss"
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
...
    UserResponse:
      type: object
      required: [id, name, email, role, status, created_at]
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        name:
          type: string
          example: "Budi Santoso"
        email:
          type: string
          format: email
          example: "budi@example.com"
        role:
          type: string
          enum: [student, contributor, admin]
        status:
          type: string
          enum: [active, inactive, suspended, pending]
        created_at:
          type: string
          format: date-time
    AuthResponse:
      type: object
      required: [access_token, refresh_token, user]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        user:
          $ref: '#/components/schemas/UserResponse'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT